requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
fuzzywuzzy==0.18.0
idna==2.10
lyricsgenius==3.0.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
spotipy==2.17.1
urllib3==1.26.4
python-dotenv~=0.17.0